==> notebooks/step02_aScrapeEspnBpi.ipynb <==
# Scrape ESPN BPI 2008 - 2022 https://www.espn.com/mens-college-basketball/bpi

from selenium import webdriver
import pandas as pd
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.common.keys import Keys
import time

PATH = 'E:\Github\chromedriver.exe' 
opt = Options()
opt.add_argument('--headless')
driver = webdriver.Chrome(PATH,options=opt)

driver.get('https://www.espn.com/mens-college-basketball/bpi/_/view/bpi/season/2008')
driver.maximize_window()

element = WebDriverWait(driver, 20).until(EC.presence_of_element_located((By.ID, "onetrust-accept-btn-handler"))).click()
time.sleep(5)
try:
    for x in range(7):
        element = WebDriverWait(driver, 10).until(EC.presence_of_element_located((By.LINK_TEXT, "Show More"))).click() 
except:
    try:
        element = WebDriverWait(driver, 1000).until(EC.presence_of_element_located((By.XPATH, "//table")))
        df = pd.read_html(driver.page_source)[0]
        df2 = pd.read_html(driver.page_source)[1]
    finally:
        
        bpi = pd.concat([df, df2], axis=1)
        bpi['Year']=2008
        bpi = bpi.iloc[:,[12,0,1,6,7]]
        bpi.columns = ['Year','Team','Conf','BPI(O)','BPI(D)']
        print(2008)
espnBPI = bpi


for y in range(2009,2023):
    driver.find_element_by_tag_name('body').send_keys(Keys.COMMAND + 't') 
    driver.get(f'https://www.espn.com/mens-college-basketball/bpi/_/season/{y}')
    time.sleep(10)
    try:
        for x in range(7):
            element = WebDriverWait(driver, 10).until(EC.presence_of_element_located((By.LINK_TEXT, "Show More"))).click() 
    except:
        try:
            element = WebDriverWait(driver, 1000).until(EC.presence_of_element_located((By.XPATH, "//table")))
            df = pd.read_html(driver.page_source)[0]
            df2 = pd.read_html(driver.page_source)[1]
        finally:
            
            bpi = pd.concat([df, df2], axis=1)
            bpi['Year']=y
            bpi = bpi.iloc[:,[12,0,1,6,7]]
            bpi.columns = ['Year','Team','Conf','BPI(O)','BPI(D)']
            print(y)
    espnBPI = pd.concat([espnBPI,bpi])  


espnBPI.to_csv('step02_espnbpi0823.csv',index=False)   
    



# Scrape ESPN BPI 2023 https://www.espn.com/mens-college-basketball/bpi

from selenium import webdriver
import pandas as pd
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
import time

PATH = 'E:\Github\chromedriver.exe' 
opt = Options()
opt.add_argument('--headless')
driver = webdriver.Chrome(PATH,options=opt)
y = 2023
driver.get(f'https://www.espn.com/mens-college-basketball/bpi/_/view/bpi/season/{y}')
driver.maximize_window()

element = WebDriverWait(driver, 20).until(EC.presence_of_element_located((By.ID, "onetrust-accept-btn-handler"))).click()
time.sleep(5)
try:
    for x in range(8):
        element = WebDriverWait(driver, 10).until(EC.presence_of_element_located((By.LINK_TEXT, "Show More"))).click() 
except:
    try:
        element = WebDriverWait(driver, 1000).until(EC.presence_of_element_located((By.XPATH, "//table")))
        df = pd.read_html(driver.page_source)[0]
        df2 = pd.read_html(driver.page_source)[1]
    finally:
        driver.quit()
        bpi = pd.concat([df, df2], axis=1)
        bpi['Year']=y
        bpi = bpi.iloc[:,[12,0,1,6,7]]
        bpi.columns = ['Year','Team','Conf','BPI(O)','BPI(D)']
espnBPI = pd.read_csv('step02_espnbpi0823.csv')
espnBPI = espnBPI[espnBPI['Year']!=y]   
espnBPI = pd.concat([espnBPI,bpi]) 
espnBPI.to_csv('step02_espnbpi0823.csv',index=False)      

a = espnBPI['Team'].unique()
b = pd.DataFrame({'tm':sorted(a)})
b.to_csv('bbb.csv',index=False)
